Don't rely on fs mapped file handles on BSD
On BSD file handles are not mapped to file handles by default, and such
mapping may not be supported on all BSD's either.

On FreeBSD we use the close_range(2) system call to set the flag on the
open handles. This function is safe to call before execve(2).

On other BSD's we simply fall back to iterating through all file handles
and set the flag individually on all open files.

This work is sponsored by The FreeBSD Foundation