Filter syms before demangling in BsdDebuggerLocal
This replicates the fix made for hotspot/decoder_bsd.cpp. The FreeBSD
implementation of abi::__cxa_demangle will return bogus data if passed
in a unmangled symbol, so we make sure to only pass it symbols that are
mangled. Otherwise we just return the symbol as is.

This work was sponsored by: The FreeBSD Foundation