Initial OpenBSD and FreeBSD github build and test actions.
aarch64 is disabled for both OpenBSD and FreeBSD because
building in an aarch64 VM running on a x86_64 host is too
slow to complete the build in under 6 hours (which is the
github timeout).

Testing for freebsd-x64 is disabled for push events because
some of the tests don't complete in under 6 hours. Testing
will run if manually starting OpenJDK GHA Sanity Checks with
freebsd-x64 in the 'Platform(s) to execute on' list.